Add repository and daos

+++ example/ddd-salesorder-domain/src/test/kotlin/io/paju/salesorder/domain/SalesOrderTestData.kt
@@ -2,6 +2,7 @@ package io.paju.salesorder.domain
 
 import io.paju.ddd.AggregateRootId
 import io.paju.ddd.EntityId
+import java.math.BigDecimal
 import java.util.UUID
 
 object SalesOrderTestData {
@@ -11,7 +12,7 @@ object SalesOrderTestData {
         return s
     }
     val customerId = EntityId.fromObject("1")
-    val product1 = Product(EntityId.fromObject(UUID.randomUUID()), Price(10.0f, Vat.vat24), "Test product1", "Test product description")
-    val product2 = Product(EntityId.fromObject(UUID.randomUUID()), Price(12.0f, Vat.vat24), "Test product2", "Test product description")
+    val product1 = Product(EntityId.fromObject(UUID.randomUUID()), Price(BigDecimal.valueOf(10.0), Vat.vat24), "Test product1", "Test product description")
+    val product2 = Product(EntityId.fromObject(UUID.randomUUID()), Price(BigDecimal.valueOf(12.0), Vat.vat24), "Test product2", "Test product description")
 }
